requirements.txt
Flask
flask-cors
Flask-Session
openai
python-dotenv